full project (add web and fix server)

+++ src/main/java/com/example/repository/DisciplinesRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 /**
  * Created by supercat on 19.4.17.
  */
-public interface DisciplinesRepository extends JpaRepository<Disciplines, Long> {
+public interface DisciplinesRepository extends JpaRepository<Disciplines, Integer> {
     Disciplines findById(int id);
     Disciplines findByNameOfDiscipline(String nameOfDiscipline);
